Use body-parser to parse POST request
body-parser package is used to parse POST request in express. This is a middleware function.
req,.body , this property has data which is to be parsed.
bodyParser.urlencoded() this function parse urlencoded bodies
bodyParser.json() this function parse only json related bodies
This middleware must be mounted before all routes which depends on it